+++ webui/react/src/components/ExperimentStopModal.tsx
@@ -1,0 +1,72 @@
+import { Alert } from 'antd';
+import { CheckboxChangeEvent } from 'antd/lib/checkbox';
+import React, { useState } from 'react';
+
+import Checkbox from 'components/kit/Checkbox';
+import { Modal } from 'components/kit/Modal';
+import { cancelExperiment, killExperiment } from 'services/api';
+import { ValueOf } from 'shared/types';
+import { ErrorLevel, ErrorType } from 'shared/utils/error';
+import handleError from 'utils/error';
+
+export const ActionType = {
+  Cancel: 'Cancel',
+  Kill: 'Kill',
+} as const;
+
+export type ActionType = ValueOf<typeof ActionType>;
+
+interface Props {
+  experimentId: number;
+  onClose?: () => void;
+}
+
+const ExperimentStopModalComponent: React.FC<Props> = ({ experimentId, onClose }: Props) => {
+  const [type, setType] = useState<ActionType>(ActionType.Cancel);
+
+  const handleCheckBoxChange = (event: CheckboxChangeEvent) => {
+    setType(event.target.checked ? ActionType.Cancel : ActionType.Kill);
+  };
+
+  const handleSubmit = async () => {
+    try {
+      if (type === ActionType.Cancel) {
+        await cancelExperiment({ experimentId });
+      } else {
+        await killExperiment({ experimentId });
+      }
+    } catch (e) {
+      handleError(e, {
+        level: ErrorLevel.Error,
+        publicMessage: 'Please try again later.',
+        publicSubject: 'Unable to stop experiment.',
+        silent: false,
+        type: ErrorType.Server,
+      });
+    }
+  };
+
+  return (
+    <Modal
+      size="small"
+      submit={{
+        handler: handleSubmit,
+        text: 'Stop Experiment',
+      }}
+      title="Confirm Stop"
+      onClose={onClose}>
+      <div>Are you sure you want to stop experiment {experimentId}?</div>
+      <Checkbox checked={type === ActionType.Cancel} onChange={handleCheckBoxChange}>
+        Save checkpoint before stopping
+      </Checkbox>
+      {type !== ActionType.Cancel && (
+        <Alert
+          message={'Note: Any progress/data on incomplete workflows will be lost.'}
+          type="warning"
+        />
+      )}
+    </Modal>
+  );
+};
+
+export default ExperimentStopModalComponent;
